+++ integration-tests/performance/ocr_test.go
@@ -27,6 +27,9 @@ import (
 func TestOCRBasic(t *testing.T) {
 	t.Parallel()
 	testEnvironment, testNetwork := setupOCRTest(t)
+	if testEnvironment.WillUseRemoteRunner() {
+		return
+	}
 
 	chainClient, err := blockchain.NewEVMClient(testNetwork, testEnvironment)
 	require.NoError(t, err, "Connecting to blockchain nodes shouldn't fail")
@@ -104,6 +107,7 @@ ListenIP = '0.0.0.0'
 ListenPort = 6690`
 	testEnvironment = environment.New(&environment.Config{
 		NamespacePrefix: fmt.Sprintf("performance-ocr-%s", strings.ReplaceAll(strings.ToLower(testNetwork.Name), " ", "-")),
+		Test:            t,
 	}).
 		AddHelm(mockservercfg.New(nil)).
 		AddHelm(mockserver.New(nil)).
